translation/requirements.txt
google-adk
google-cloud-translate>=3.11.0
google-cloud-storage
pydantic
pypdf2
python-dotenv
